Refactoring pycmake

+++ test/print_0_5/main.cpp
@@ -4,6 +4,6 @@ int main()
 {
     for (size_t i = 0; i <= 5; i++)
     {
-        PrintTest(i);
+        PrintTest((double)i);
     }
 }
